+++ content-src/components/ActivityFeed/ActivityFeed.js
@@ -3,6 +3,7 @@ const {connect} = require("react-redux");
 const {justDispatch} = require("selectors/selectors");
 const {actions} = require("common/action-manager");
 const SiteIcon = require("components/SiteIcon/SiteIcon");
+const DeleteMenu = require("components/DeleteMenu/DeleteMenu");
 const {prettyUrl, getRandomFromTimestamp} = require("lib/utils");
 const moment = require("moment");
 const classNames = require("classnames");
@@ -10,15 +11,30 @@ const classNames = require("classnames");
 const ICON_SIZE = 16;
 const TOP_LEFT_ICON_SIZE = 20;
 const SESSION_DIFF = 600000;
+const CALENDAR_HEADINGS = {
+  sameDay: "[Today]",
+  nextDay: "[Tomorrow]",
+  nextWeek: "dddd",
+  lastDay: "[Yesterday]",
+  lastWeek: "[Last] dddd",
+  sameElse: "DD/MM/YYYY"
+};
 
 const ActivityFeedItem = React.createClass({
+  getInitialState() {
+    return {
+      showContextMenu: false
+    };
+  },
   getDefaultProps() {
     return {
       onShare: function() {},
       onClick: function() {},
-      onDelete: function() {},
       showDate: false
     };
+  },
+  onDeleteClick() {
+    this.setState({showContextMenu: true});
   },
   render() {
     const site = this.props;
@@ -47,7 +63,7 @@ const ActivityFeedItem = React.createClass({
       dateLabel = moment(date).format("h:mm A");
     }
 
-    return (<li className={classNames("feed-item", {bookmark: site.bookmarkGuid})}>
+    return (<li className={classNames("feed-item", {bookmark: site.bookmarkGuid, fixed: this.state.showContextMenu})}>
       <a onClick={this.props.onClick} href={site.url} ref="link">
         <span className="star" hidden={!site.bookmarkGuid} />
         {icon}
@@ -62,15 +78,26 @@ const ActivityFeedItem = React.createClass({
         </div>
       </a>
       <div className="action-items-container">
-        <div className="action-item icon-delete" ref="delete" onClick={() => this.props.onDelete(site.url)}></div>
+        <div className="action-item icon-delete" ref="delete" onClick={this.onDeleteClick}></div>
         <div className="action-item icon-share" ref="share" onClick={() => this.props.onShare(site.url)}></div>
         <div className="action-item icon-more" onClick={() => alert("Sorry. We are still working on this feature.")}></div>
       </div>
+      <DeleteMenu
+        visible={this.state.showContextMenu}
+        onUpdate={val => this.setState({showContextMenu: val})}
+        url={site.url}
+        page={this.props.page}
+        index={this.props.index}
+        source={this.props.source}
+        />
     </li>);
   }
 });
 
 ActivityFeedItem.propTypes = {
+  page: React.PropTypes.string,
+  source: React.PropTypes.string,
+  index: React.PropTypes.number,
   onShare: React.PropTypes.func,
   onClick: React.PropTypes.func,
   url: React.PropTypes.string.isRequired,
@@ -122,24 +149,14 @@ function groupSitesBySession(sites) {
 const GroupedActivityFeed = React.createClass({
   getDefaultProps() {
     return {
-      dateKey: "lastVisitDate"
+      dateKey: "lastVisitDate",
+      showDateHeadings: false
     };
   },
   onClickFactory(index) {
     return () => {
       this.props.dispatch(actions.NotifyEvent({
         event: "CLICK",
-        page: this.props.page,
-        source: "ACTIVITY_FEED",
-        action_position: index
-      }));
-    };
-  },
-  onDeleteFactory(index) {
-    return url => {
-      this.props.dispatch(actions.NotifyHistoryDelete(url));
-      this.props.dispatch(actions.NotifyEvent({
-        event: "DELETE",
         page: this.props.page,
         source: "ACTIVITY_FEED",
         action_position: index
@@ -167,10 +184,13 @@ const GroupedActivityFeed = React.createClass({
     let globalCount = -1;
     return (<div className="grouped-activity-feed">
       {sites.length > 0 && this.props.title &&
-        <h3 className="section-title">{this.props.title}</h3>
+        <h3 ref="title" className="section-title">{this.props.title}</h3>
       }
-      {Array.from(groupedSites.keys()).map(date => {
-        return (<div key={date}>
+      {Array.from(groupedSites.keys()).map((date, dateIndex) => {
+        return (<div className="group" key={date}>
+          {this.props.showDateHeadings && dateIndex > 0 &&
+            <h3 className="section-title">{moment(date).startOf("day").calendar(null, CALENDAR_HEADINGS)}</h3>
+          }
           {groupedSites.get(date).map((sites, outerIndex) => {
             return (<ul key={date + "-" + outerIndex} className="activity-feed">
               {sites.map((site, i) => {
@@ -179,9 +199,11 @@ const GroupedActivityFeed = React.createClass({
                     key={site[this.props.dateKey]}
                     onClick={this.onClickFactory(globalCount)}
                     onShare={this.onShareFactory(globalCount)}
-                    onDelete={this.onDeleteFactory(globalCount)}
                     showImage={getRandomFromTimestamp(0.2, site)}
-                    showDate={i === 0}
+                    index={globalCount}
+                    page={this.props.page}
+                    source="ACTIVITY_FEED"
+                    showDate={!this.props.showDateHeadings && outerIndex === 0 && i === 0}
                     {...site} />);
               })}
             </ul>);
@@ -197,7 +219,8 @@ GroupedActivityFeed.propTypes = {
   length: React.PropTypes.number,
   title: React.PropTypes.string,
   dateKey: React.PropTypes.string,
-  page: React.PropTypes.string
+  page: React.PropTypes.string,
+  showDateHeadings: React.PropTypes.bool
 };
 
 module.exports = connect(justDispatch)(GroupedActivityFeed);