+++ content-src/selectors/selectors.js
@@ -7,26 +7,57 @@ const urlParse = require("url-parse");
 const SPOTLIGHT_LENGTH = module.exports.SPOTLIGHT_LENGTH = 3;
 const TOP_SITES_LENGTH = module.exports.TOP_SITES_LENGTH = 6;
 const BACKGROUND_FADE = 0.5;
-const DEFAULT_FAVICON_BG_COLOR = [255, 255, 255];
+const DEFAULT_FAVICON_BG_COLOR = [150, 150, 150];
 
 module.exports.justDispatch = (() => ({}));
 
+function getBackgroundRGB(site) {
+  if (site.favicon_colors && site.favicon_colors[0] && site.favicon_colors[0].color) {
+    return site.favicon_colors[0].color;
+  }
+
+  const favicon = site.favicon_url || site.favicon;
+  const parsedUrl = site.parsedUrl || urlParse(site.url || "") ;
+  const label = prettyUrl(parsedUrl.hostname);
+  return favicon ? DEFAULT_FAVICON_BG_COLOR : getRandomColor(label);
+}
+
+module.exports.getBackgroundRGB = getBackgroundRGB;
+
+function isValidSpotlightSite(site) {
+  return (site.bestImage &&
+    site.title &&
+    site.description &&
+    site.title !== site.description);
+}
+
 const selectSpotlight = module.exports.selectSpotlight = createSelector(
-  [state => state.FrecentHistory, state => state.Blocked],
+  [
+    state => state.FrecentHistory,
+    state => state.Blocked
+  ],
   (FrecentHistory, Blocked) => {
     const rows = FrecentHistory.rows
     .filter(site => !Blocked.urls.has(site.url))
     .map(site => {
+      const newProps = {};
       const bestImage = getBestImage(site.images);
-      return Object.assign({}, site, {bestImage});
+      if (bestImage) {
+        newProps.bestImage = bestImage;
+      }
+      newProps.backgroundColor = toRGBString(...getBackgroundRGB(site, [200, 200, 200]), BACKGROUND_FADE - 0.1);
+      return Object.assign({}, site, newProps);
     })
-    .filter(site => {
-      return (
-        site.bestImage &&
-        site.title &&
-        site.description &&
-        site.title !== site.description
-      );
+    .sort((site1, site2) => {
+      const site1Valid = isValidSpotlightSite(site1);
+      const site2Valid = isValidSpotlightSite(site2);
+      if (site1Valid && site2Valid) {
+        return 0;
+      } else if (site2Valid) {
+        return 1;
+      } else {
+        return -1;
+      }
     });
     return Object.assign({}, FrecentHistory, {rows});
   }
@@ -37,10 +68,11 @@ module.exports.selectNewTabSites = createSelector(
     state => state.TopSites,
     state => state.FrecentHistory,
     state => state.History,
+    state => state.Highlights,
     selectSpotlight,
     state => state.Blocked
   ],
-  (TopSites, FrecentHistory, History, Spotlight, Blocked) => {
+  (TopSites, FrecentHistory, History, Highlights, Spotlight, Blocked) => {
 
     // Removed blocked
     [TopSites, Spotlight] = [TopSites, Spotlight].map(item => {
@@ -56,27 +88,26 @@ module.exports.selectNewTabSites = createSelector(
     // Dedupe top activity
     const topActivityRows = dedupe.group([topSitesRows, spotlightRows, History.rows])[2].sort((a, b) => {
       return b.lastVisitDate - a.lastVisitDate;
-    });
+    }).filter(site => !Blocked.urls.has(site.url));
 
     return {
       TopSites: Object.assign({}, TopSites, {rows: topSitesRows}),
       Spotlight: Object.assign({}, FrecentHistory, {rows: spotlightRows}),
       TopActivity: Object.assign({}, History, {rows: topActivityRows}),
-      isReady: TopSites.init && FrecentHistory.init && History.init
+      isReady: TopSites.init && FrecentHistory.init && History.init && Highlights.init
     };
   }
 );
 
-module.exports.selectSiteIcon = createSelector(
+const selectSiteIcon = createSelector(
   site => site,
   site => {
     const favicon = site.favicon_url || site.favicon;
     const parsedUrl = site.parsedUrl || urlParse(site.url || "") ;
     const label = prettyUrl(parsedUrl.hostname);
-    let background = favicon ? DEFAULT_FAVICON_BG_COLOR : getRandomColor(label);
-    try { background = site.favicon_colors[0].color || background; } catch (e) {/**/}
-    const backgroundColor = toRGBString(...background, favicon ? BACKGROUND_FADE : 1);
-    const fontColor = getBlackOrWhite(...background);
+    const backgroundRGB = getBackgroundRGB(site);
+    const backgroundColor = toRGBString(...backgroundRGB, favicon ? BACKGROUND_FADE : 1);
+    const fontColor = getBlackOrWhite(...backgroundRGB);
     const firstLetter = prettyUrl(parsedUrl.hostname)[0];
     return {
       url: site.url,
@@ -89,4 +120,34 @@ module.exports.selectSiteIcon = createSelector(
   }
 );
 
+// Timeline History view
+module.exports.selectHistory = createSelector(
+  [
+    selectSpotlight,
+    state => state.History,
+    state => state.Blocked
+  ],
+  (Spotlight, History, Blocked) => {
+    return {
+      Spotlight,
+      History: Object.assign({}, History, {rows: History.rows.filter(site => !Blocked.urls.has(site.url))})
+    };
+  }
+);
+
+// Timeline Bookmarks
+module.exports.selectBookmarks = createSelector(
+  [
+    state => state.Bookmarks,
+    state => state.Blocked
+  ],
+  (Bookmarks, Blocked) => {
+    return {
+      Bookmarks: Object.assign({}, Bookmarks, {rows: Bookmarks.rows.filter(site => !Blocked.urls.has(site.url))})
+    };
+  }
+);
+
+module.exports.selectSiteIcon = selectSiteIcon;
 module.exports.selectSiteIcon.BACKGROUND_FADE = BACKGROUND_FADE;
+module.exports.DEFAULT_FAVICON_BG_COLOR = DEFAULT_FAVICON_BG_COLOR;