+++ content-test/components/ActivityFeed.test.js
@@ -1,12 +1,14 @@
 const {assert} = require("chai");
 const ConnectedActivityFeed = require("components/ActivityFeed/ActivityFeed");
 const {ActivityFeedItem, GroupedActivityFeed, groupSitesBySession} = ConnectedActivityFeed;
+const DeleteMenu = require("components/DeleteMenu/DeleteMenu");
 const SiteIcon = require("components/SiteIcon/SiteIcon");
 const React = require("react");
 const ReactDOM = require("react-dom");
 const TestUtils = require("react-addons-test-utils");
 const {prettyUrl} = require("lib/utils");
 const moment = require("moment");
+const {renderWithProvider} = require("test/test-utils");
 
 const {faker, overrideConsoleError} = require("test/test-utils");
 const fakeSite = {
@@ -33,14 +35,14 @@ describe("ActivityFeedItem", function() {
   let instance;
   let el;
   beforeEach(() => {
-    instance = TestUtils.renderIntoDocument(<ActivityFeedItem {...fakeSite} />);
+    instance = renderWithProvider(<ActivityFeedItem {...fakeSite} />);
     el = ReactDOM.findDOMNode(instance);
   });
 
   it("should not throw if missing props", () => {
     assert.doesNotThrow(() => {
       const restore = overrideConsoleError();
-      TestUtils.renderIntoDocument(<ActivityFeedItem />);
+      renderWithProvider(<ActivityFeedItem />);
       restore();
     });
   });
@@ -58,36 +60,34 @@ describe("ActivityFeedItem", function() {
       assert.include(instance.refs.icon.props.site, fakeSite);
     });
     it("should render the title", () => {
-      assert.equal(instance.refs.title.innerHTML, fakeSite.title);
+      assert.equal(instance.refs.title.textContent, fakeSite.title);
     });
     it("should render the url", () => {
       const urlEl = instance.refs.url;
-      assert.equal(urlEl.innerHTML, prettyUrl(fakeSite.url));
+      assert.equal(urlEl.textContent, prettyUrl(fakeSite.url));
     });
     it("should render the time", () => {
       const lastVisitEl = instance.refs.lastVisit;
-      assert.equal(lastVisitEl.innerHTML, moment(fakeSite.dateDisplay).format("h:mm A"));
+      assert.equal(lastVisitEl.textContent, moment(fakeSite.dateDisplay).format("h:mm A"));
     });
     it("should not have a bookmark class if no bookmarkGuid", () => {
       assert.notInclude(el.className, "bookmark");
     });
     it("should have a bookmark class if the site has a bookmarkGuid", () => {
-      instance = TestUtils.renderIntoDocument(<ActivityFeedItem {...fakeSiteWithBookmark} />);
+      instance = renderWithProvider(<ActivityFeedItem {...fakeSiteWithBookmark} />);
       assert.include(ReactDOM.findDOMNode(instance).className, "bookmark");
     });
-    it("should call onDelete callback with url when delete icon is pressed", done => {
-      function onDelete(url) {
-        assert.equal(url, fakeSite.url);
-        done();
-      }
-      const item = TestUtils.renderIntoDocument(<ActivityFeedItem onDelete={onDelete} {...fakeSite} />);
+    it("should show the delete menu when the delete button is clicked", () => {
+      const item = renderWithProvider(<ActivityFeedItem {...fakeSite} />);
       const button = item.refs.delete;
       TestUtils.Simulate.click(button);
+      const deleteMenu = TestUtils.findRenderedComponentWithType(item, DeleteMenu);
+      assert.equal(deleteMenu.props.visible, true);
     });
     it("should render date if showDate=true", () => {
-      const item = TestUtils.renderIntoDocument(<ActivityFeedItem showDate={true} {...fakeSite} />);
+      const item = renderWithProvider(<ActivityFeedItem showDate={true} {...fakeSite} />);
       const lastVisitEl = item.refs.lastVisit;
-      assert.equal(lastVisitEl.innerHTML, moment(fakeSite.dateDisplay).calendar());
+      assert.equal(lastVisitEl.textContent, moment(fakeSite.dateDisplay).calendar());
     });
   });
 });
@@ -102,7 +102,7 @@ describe("GroupedActivityFeed", function() {
       faker.createSite({moment: faker.moment().subtract(2, "days")}),
       faker.createSite({moment: faker.moment().subtract(4, "days")}),
     ];
-    instance = TestUtils.renderIntoDocument(<GroupedActivityFeed sites={sites} />);
+    instance = renderWithProvider(<GroupedActivityFeed sites={sites} />);
     el = ReactDOM.findDOMNode(instance);
   });
 
@@ -117,8 +117,38 @@ describe("GroupedActivityFeed", function() {
       assert.equal(children.length, sites.length);
     });
     it("shouldn't render title if there are no sites", () => {
-      const item = TestUtils.renderIntoDocument(<GroupedActivityFeed sites={[]} title="Fake Title" />);
-      assert.isNull(ReactDOM.findDOMNode(item).querySelector(".section-title"));
+      const item = renderWithProvider(<GroupedActivityFeed sites={[]} title="Fake Title" />);
+      assert.isUndefined(item.refs.title);
+    });
+  });
+
+  describe("date headings", () => {
+    let m;
+    let m2;
+    let m3;
+    let sites;
+    beforeEach(() => {
+      const date = new Date();
+      m = moment(date);
+      m2 = moment(date).subtract(1, "days");
+      m3 = moment(date).subtract(2, "days");
+      sites = [
+        faker.createSite({moment: m}),
+        faker.createSite({moment: m2}),
+        faker.createSite({moment: m3})
+      ];
+    });
+    it("should show date headings if showDateHeadings is true", () => {
+      const item = renderWithProvider(<GroupedActivityFeed showDateHeadings={true} sites={sites} />);
+      const titles = TestUtils.scryRenderedDOMComponentsWithClass(item, "section-title");
+      assert.lengthOf(titles, 2);
+      assert.equal(titles[0].innerHTML, "Yesterday");
+      assert.equal(titles[1].innerHTML, m3.format("[Last] dddd"));
+    });
+    it("should not show date headings if showDateHeadings is false", () => {
+      const item = renderWithProvider(<GroupedActivityFeed showDateHeadings={false} sites={sites} />);
+      const titles = TestUtils.scryRenderedDOMComponentsWithClass(item, "section-title");
+      assert.lengthOf(titles, 0);
     });
   });
 
@@ -132,22 +162,8 @@ describe("GroupedActivityFeed", function() {
           done();
         }
       }
-      instance = TestUtils.renderIntoDocument(<GroupedActivityFeed dispatch={dispatch} page={"NEW_TAB"} sites={sites} />);
+      instance = renderWithProvider(<GroupedActivityFeed dispatch={dispatch} page={"NEW_TAB"} sites={sites} />);
       const link = TestUtils.scryRenderedComponentsWithType(instance, ActivityFeedItem)[0].refs.link;
-      TestUtils.Simulate.click(link);
-    });
-    it("should send an event onDelete", done => {
-      function dispatch(a) {
-        if (a.type === "NOTIFY_USER_EVENT") {
-          assert.equal(a.data.event, "DELETE");
-          assert.equal(a.data.page, "NEW_TAB");
-          assert.equal(a.data.source, "ACTIVITY_FEED");
-          assert.equal(a.data.action_position, 1);
-          done();
-        }
-      }
-      instance = TestUtils.renderIntoDocument(<GroupedActivityFeed dispatch={dispatch} page={"NEW_TAB"} sites={sites} />);
-      const link = TestUtils.scryRenderedComponentsWithType(instance, ActivityFeedItem)[1].refs.delete;
       TestUtils.Simulate.click(link);
     });
     it("should send an event onShare", done => {
@@ -160,7 +176,7 @@ describe("GroupedActivityFeed", function() {
           done();
         }
       }
-      instance = TestUtils.renderIntoDocument(<GroupedActivityFeed dispatch={dispatch} page={"NEW_TAB"} sites={sites} />);
+      instance = renderWithProvider(<GroupedActivityFeed dispatch={dispatch} page={"NEW_TAB"} sites={sites} />);
       const link = TestUtils.scryRenderedComponentsWithType(instance, ActivityFeedItem)[2].refs.share;
       TestUtils.Simulate.click(link);
     });